66강-2: Json은 Long을 표현할 수 없다. 따라서 memberId를 Long으로 살리는 작업 추가

+++ src/main/java/com/rest1/global/rq/Rq.java
@@ -49,7 +49,7 @@ public class Rq {
             Map<String, Object> payload = memberService.payloadOrNull(accessToken);
 
             if (payload != null) {
-                int id = (int) payload.get("id");
+                long id = (long) payload.get("id");
                 member = memberService.findById(id)
                         .orElseThrow(() -> new ServiceException("401-4", "accessToken의 id에 해당하는 회원이 존재하지 않습니다."));
             }
